Extract changed-files detector
