+++ .buildkite/scripts/flakiness-detection/detectors/changed-files.ts
@@ -1,0 +1,34 @@
+import {
+  ClassifiedTest,
+  SOURCE_SET_PATTERNS,
+  toFqcn,
+  toGradleProject,
+} from "../domain";
+
+export function classifyChangedFiles(files: string[]): ClassifiedTest[] {
+  const tests: ClassifiedTest[] = [];
+
+  for (const file of files) {
+    for (const pattern of SOURCE_SET_PATTERNS) {
+      const match = file.match(pattern.regex);
+      if (match) {
+        const test: ClassifiedTest = {
+          gradleProject: toGradleProject(match[1]),
+          kind: pattern.kind,
+          sourceSet: pattern.sourceSet,
+        };
+
+        if (pattern.kind === "yamlRestTestSuite") {
+          test.suitePath = match[2];
+        } else if (pattern.kind !== "yamlRestTestRunner") {
+          test.fqcn = toFqcn(match[2]);
+        }
+
+        tests.push(test);
+        break;
+      }
+    }
+  }
+
+  return tests;
+}
